function_resolver/requirements.txt
fastapi
sentence-transformers
torch[mps]
uvicorn
